Don't ignore dist directory anymore
The use of the dist rule is to create distribution files for the package. In so
doing, the procedure will be to create a tagged branch in which we actually add
the dist files. So it's no longer necessary or desired to ignore "dist" - in
fact, it could lead to problems.