Removed implicit render() call from all component constructors
The call doesn't seem to be necessary. In the examples, for instance, the chart
is rendered with an explicit render() call. This is the right thing to do in any
case, because there are times when you will want to construct the chart but not
render it, and our previous approach made it impossible to accomplish this.

Furthermore, this enables a simpler approach to mixins. When using the Events
mixin, the call to super() (to invoke the parent constructor) was previously
causing render() to be called before the mixin has a chance to mix in. That
means in that one initial call to render(), the Events mixin not having yet been
invoked, the call to this.emit() fails somewhat mysteriously. The (ugly)
workaround was to wrap the call to this.emit() in an if(this.emit) block.