Don't sort on time in the spec
This is apparently an unstable sort, so we avoid it. The data is already sorted
by the component constructor in any case. Sorting again here changes the
relative order of the data points, which can disrupt the ordering of the
"survivors" value. This, in turn, can mess up the order in which the stepped
lines are drawn.