feat(ScatterPlot): add opacity option to ScatterPlot
Add a static "opacity" option to ScatterPlot's config object. This represents a
CSS opacity value for the "fill" property of the symbols used by the ScatterPlot
to render data. The default value of 0.01 leaves the symbols effectively
transparent while still allowing them to intercept pointer events. This option
allows for solid symbols as well.