ci: swallow errors from `npm run semantic-release`
This is done because technically, semantic-release fails with non-zero
exit code on ENOCHANGE, which is the error code generated when there is
nothing to actually release (i.e., if there are no `feat` or `fix`
commit tags).

Travis gets around this by not failing the entire build if the
`after_success` step fails; in Circle we're using the deployment step
for this, which *does* fail the whole build if it fails.