build: move "runtime" dependencies into "dependencies" field
"Runtime" appears in quotes because these are not simply dependencies
that are needed when Candela is used - it includes dependencies that
*may be* needed when Candela is used *to build a client project*. Since
Candela takes the form of a library, but one that supplies its raw
source code for use in downstream projects, it must include certain
things as "dependencies" that are perhaps more naturally
"devDependencies". This includes loaders for webpack for processing ES6
code, etc.