docs(candela): use blank strings instead of None for version/release values
Sphinx 1.5 chokes when these values are specified as `None`, but using
blank strings is fine too.

The strings appear in-page as a version string, but RTD places the
version at the bottom of the menu bar in any case, so we don't need to
put anything here. It would be rather difficult in any case to determine
what version string should go here for automatic builds, etc.